requirements.txt
RPi.GPIO
click
pyyaml
requests
simple-pid
systemd-python
w1thermsensor
